move browser, default, environment credentials
